Move Resources to pages
